Fix bugs in the creation of interviews and answers
QuestionSets were being created when an interview was created, yielding empty qs objects for new interviews. Changed the new interview page to submit just the interview data, not the full nested QS form, this makes is trivial to get the ID of the added interview for the forward. Also, whenever an interview was conducted, it was creating all the answers, even if they already existed, leading to dupes. Now it assumes that if there are any answers, then it creates none.